% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multiserver.R
\name{Multiserver}
\alias{Multiserver}
\title{Multiserver function}
\usage{
Multiserver(Arrivals, ServiceTimes, NumServers = 1)
}
\arguments{
\item{Arrivals}{vector of quantities.  Which detail the arrival time of a customer.}

\item{ServiceTimes}{vector of quantities.  Which detail how long it takes for a customer's transaction to finish.}

\item{NumServers}{number of servers.}
}
\value{
A data frame.
}
\description{
Simulates how customers will go through a first come, first served queuing system.  Given that "fate has already decided when each customer arrives and their service timess.
}
\examples{
Multiserver(10, 5, 1)
Multiserver(c(1,20), c(5,12), 2)


}
